backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
chromadb==0.5.5
openai==1.47.0
python-dotenv==1.0.1
reportlab==4.2.2
PyPDF2==3.0.1
pdfplumber==0.11.4
sentence-transformers==3.1.0
scikit-learn==1.5.2
numpy==1.26.4
aiofiles==24.1.0
